Disable overtriggered lint check for from-text stub build
ObjectToString lint check is triggered when a final class does not
override Object.ToString(). Per Android API Guidelines, this is an
encouraged behavior but not required. Also, since the stubs generated
from signature files do not include overriding methods with identical
signatures to its parent method(s), majority of the toString() method
will not be included in the stubs, triggering this lint warning.
Therefore, this change disables the ObjectToString error prone check.

Test: m droid
Bug: 306242755
Change-Id: Iaf4a7f62f33e5cc3fb024e346d95ad3a0ee81e08